fixup revert babel config move
